Split layout code into a separate crate.
