+++ src/components/layout/construct.rs
@@ -20,31 +20,31 @@
 #![deny(unsafe_block)]
 
 use css::node_style::StyledNode;
-use layout::block::BlockFlow;
-use layout::context::LayoutContext;
-use layout::floats::FloatKind;
-use layout::flow::{Flow, ImmutableFlowUtils, MutableOwnedFlowUtils};
-use layout::flow::{Descendants, AbsDescendants};
-use layout::flow;
-use layout::flow_ref::FlowRef;
-use layout::fragment::{Fragment, GenericFragment, IframeFragment, IframeFragmentInfo};
-use layout::fragment::{ImageFragment, ImageFragmentInfo, SpecificFragmentInfo, TableFragment};
-use layout::fragment::{TableCellFragment, TableColumnFragment, TableColumnFragmentInfo};
-use layout::fragment::{TableRowFragment, TableWrapperFragment, UnscannedTextFragment};
-use layout::fragment::{UnscannedTextFragmentInfo};
-use layout::inline::{FragmentIndex, InlineFragments, InlineFlow};
-use layout::parallel;
-use layout::table_wrapper::TableWrapperFlow;
-use layout::table::TableFlow;
-use layout::table_caption::TableCaptionFlow;
-use layout::table_colgroup::TableColGroupFlow;
-use layout::table_rowgroup::TableRowGroupFlow;
-use layout::table_row::TableRowFlow;
-use layout::table_cell::TableCellFlow;
-use layout::text::TextRunScanner;
-use layout::util::{LayoutDataAccess, OpaqueNodeMethods};
-use layout::wrapper::{PostorderNodeMutTraversal, TLayoutNode, ThreadSafeLayoutNode};
-use layout::wrapper::{Before, BeforeBlock, After, AfterBlock, Normal};
+use block::BlockFlow;
+use context::LayoutContext;
+use floats::FloatKind;
+use flow::{Flow, ImmutableFlowUtils, MutableOwnedFlowUtils};
+use flow::{Descendants, AbsDescendants};
+use flow;
+use flow_ref::FlowRef;
+use fragment::{Fragment, GenericFragment, IframeFragment, IframeFragmentInfo};
+use fragment::{ImageFragment, ImageFragmentInfo, SpecificFragmentInfo, TableFragment};
+use fragment::{TableCellFragment, TableColumnFragment, TableColumnFragmentInfo};
+use fragment::{TableRowFragment, TableWrapperFragment, UnscannedTextFragment};
+use fragment::{UnscannedTextFragmentInfo};
+use inline::{FragmentIndex, InlineFragments, InlineFlow};
+use parallel;
+use table_wrapper::TableWrapperFlow;
+use table::TableFlow;
+use table_caption::TableCaptionFlow;
+use table_colgroup::TableColGroupFlow;
+use table_rowgroup::TableRowGroupFlow;
+use table_row::TableRowFlow;
+use table_cell::TableCellFlow;
+use text::TextRunScanner;
+use util::{LayoutDataAccess, OpaqueNodeMethods};
+use wrapper::{PostorderNodeMutTraversal, TLayoutNode, ThreadSafeLayoutNode};
+use wrapper::{Before, BeforeBlock, After, AfterBlock, Normal};
 
 use gfx::display_list::OpaqueNode;
 use gfx::font_context::FontContext;
