+++ src/components/layout/model.rs
@@ -6,7 +6,7 @@
 
 #![deny(unsafe_block)]
 
-use layout::fragment::Fragment;
+use fragment::Fragment;
 
 use computed = style::computed_values;
 use geom::SideOffsets2D;
