+++ src/components/layout/table_rowgroup.rs
@@ -6,15 +6,15 @@
 
 #![deny(unsafe_block)]
 
-use layout::block::BlockFlow;
-use layout::block::WidthAndMarginsComputer;
-use layout::construct::FlowConstructor;
-use layout::context::LayoutContext;
-use layout::flow::{TableRowGroupFlowClass, FlowClass, Flow, ImmutableFlowUtils};
-use layout::flow;
-use layout::fragment::Fragment;
-use layout::table::{InternalTable, TableFlow};
-use layout::wrapper::ThreadSafeLayoutNode;
+use block::BlockFlow;
+use block::WidthAndMarginsComputer;
+use construct::FlowConstructor;
+use context::LayoutContext;
+use flow::{TableRowGroupFlowClass, FlowClass, Flow, ImmutableFlowUtils};
+use flow;
+use fragment::Fragment;
+use table::{InternalTable, TableFlow};
+use wrapper::ThreadSafeLayoutNode;
 
 use servo_util::geometry::Au;
 use servo_util::geometry;
